+++ TrainingMasterWebAPI/Models/EF/foodItem.cs
@@ -22,7 +22,7 @@ namespace TrainingMasterWebAPI.Models.EF
     
         public int FIID { get; set; }
         public string name { get; set; }
-        public int kcal { get; set; }
+        public decimal kcal { get; set; }
         public decimal fat { get; set; }
         public decimal saturatedFat { get; set; }
         public decimal unsaturatedFat { get; set; }
